# swagger.yaml
openapi: 3.0.0
info:
  title: Car Contract Management System API
  description: 중고차 계약 관리 시스템 API 문서
  version: 1.0.0
  contact:
    name: API Support
    email: support@carmate.com
servers:
  - url: http://localhost:${process.env.PORT || 3000}
    description: Development server
    # 배포 전이므로 주석 처리
  # - url: ${process.env.PROD_SERVER_URL || 'https://api.carmate.com'}
  #  description: Production server
...
security:
  - bearerAuth: []
paths:
...
  /customers:
    post:
      tags:
        - Customers
      summary: 고객 등록
      description: 새로운 고객 정보 등록
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomerRequest'
      responses:
        '201':
          description: 고객 등록 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

    get:
      tags:
        - Customers
      summary: 고객 목록 조회
      description: 등록된 고객 목록 조회
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - in: query
          name: searchBy
          schema:
            type: string
            enum: [name, email]
        - $ref: '#/components/parameters/KeywordParam'
      responses:
        '200':
          description: 고객 목록 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /customers/{customerId}:
    get:
      tags:
        - Customers
      summary: 고객 상세 조회
      description: 특정 고객의 상세 정보 조회
      parameters:
        - in: path
          name: customerId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 고객 상세 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - Customers
      summary: 고객 정보 수정
      description: 고객 정보 수정
      parameters:
        - in: path
          name: customerId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCustomerRequest'
      responses:
        '200':
          description: 고객 정보 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Customers
      summary: 고객 삭제
      description: 고객 정보 삭제
      parameters:
        - in: path
          name: customerId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 고객 삭제 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: '고객 삭제 성공'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /customers/upload:
    post:
      tags:
        - Customers
      summary: 고객 대용량 업로드
      description: CSV 파일을 통한 고객 정보 대용량 등록
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: 고객 업로드 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: '업로드가 완료되었습니다.'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      in: query
      name: page
      schema:
        type: integer
        default: 1
      description: 페이지 번호
    PageSizeParam:
      in: query
      name: pageSize
      schema:
        type: integer
        default: 8
      description: 페이지당 아이템 수
    KeywordParam:
      in: query
      name: keyword
      schema:
        type: string
      description: 검색 키워드
  schemas:
...
    CreateCustomerRequest:
      type: object
      required:
        - name
        - gender
        - phoneNumber
        - email
      properties:
        name:
          type: string
        gender:
          $ref: '#/components/schemas/Gender'
        phoneNumber:
          type: string
        ageGroup:
          $ref: '#/components/schemas/AgeGroup'
        region:
          $ref: '#/components/schemas/Region'
        email:
          type: string
          format: email
        memo:
          type: string
    UpdateCustomerRequest:
      type: object
      properties:
        name:
          type: string
        gender:
          $ref: '#/components/schemas/Gender'
        phoneNumber:
          type: string
        ageGroup:
          $ref: '#/components/schemas/AgeGroup'
        region:
          $ref: '#/components/schemas/Region'
        email:
          type: string
          format: email
        memo:
          type: string
    CustomerResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        gender:
          $ref: '#/components/schemas/Gender'
        phoneNumber:
          type: string
        ageGroup:
          $ref: '#/components/schemas/AgeGroup'
        region:
          $ref: '#/components/schemas/Region'
        email:
          type: string
        memo:
          type: string
        contractCount:
          type: integer
    CustomerListResponse:
      type: object
      properties:
        currentPage:
          type: integer
        totalPages:
          type: integer
        totalItemCount:
          type: integer
        customers:
          type: array
          items:
            $ref: '#/components/schemas/CustomerResponse'

    # Contract Document schemas
...
    Gender:
      type: string
      enum:
        - male
        - female
    AgeGroup:
      type: string
      enum:
        - 10대
        - 20대
        - 30대
        - 40대
        - 50대
        - 60대
        - 70대
        - 80대
    Region:
      type: string
      enum:
        - 서울
        - 경기
        - 인천
        - 강원
        - 충북
        - 충남
        - 세종
        - 대전
        - 전북
        - 전남
        - 광주
        - 경북
        - 경남
        - 대구
        - 울산
        - 부산
        - 제주
...
    Error:
      type: object
      properties:
        message:
          type: string
        code:
          type: string
        timestamp:
          type: string
          format: date-time
  responses:
    BadRequest:
      description: 잘못된 요청
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: '잘못된 요청입니다'
            code: 'BAD_REQUEST'
            timestamp: '2024-01-01T00:00:00Z'
    Unauthorized:
      description: 인증이 필요합니다
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: '로그인이 필요합니다'
            code: 'UNAUTHORIZED'
            timestamp: '2024-01-01T00:00:00Z'
...
    NotFound:
      description: 리소스를 찾을 수 없습니다
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: '존재하지 않는 리소스입니다'
            code: 'NOT_FOUND'
            timestamp: '2024-01-01T00:00:00Z'
...